+++ lib/knock.rb
@@ -26,6 +26,13 @@ module Knock
   mattr_accessor :token_public_key
   self.token_public_key = nil
 
+  mattr_accessor :not_found_exception_class_name
+  self.not_found_exception_class_name = 'ActiveRecord::RecordNotFound'
+
+  def self.not_found_exception_class
+    not_found_exception_class_name.to_s.constantize
+  end
+
   # Default way to setup Knock. Run `rails generate knock:install` to create
   # a fresh initializer with all configuration values.
   def self.setup