+++ test/dummy/config/initializers/knock.rb
@@ -1,0 +1,10 @@
+Knock.setup do |config|
+  config.token_signature_algorithm = 'HS256'
+  config.token_secret_signature_key = -> { Rails.application.secrets.secret_key_base }
+  config.token_public_key = nil
+  config.token_audience = nil
+
+  config.current_user_from_handle = -> handle { User.find_by(Knock.handle_attr => handle) || raise(Knock::MyCustomException) }
+  config.current_user_from_token = -> claims { User.find_by(id: claims['sub']) || raise(Knock::MyCustomException) }
+  config.not_found_exception_class_name = 'Knock::MyCustomException'
+end